requirements.txt
# Logística de Caminhões - dependências
# Para uso local ou Databricks (Databricks já inclui PySpark e Delta)

pyspark>=3.4.0
delta-spark>=2.4.0
faker>=22.0.0

# Deploy de dashboard (script scripts/deploy_dashboard.py)
databricks-sdk>=0.40.0

# Lakebase (PostgreSQL)
psycopg2-binary>=2.9.0
